# Cherry-pick #16587 to 7.x: Reduce memory allocations in decode_cef processor