--- a/x-pack/filebeat/processors/decode_cef/cef/cmd/cef2json/cef2json.go
+++ b/x-pack/filebeat/processors/decode_cef/cef/cmd/cef2json/cef2json.go
@@ -49,7 +49,7 @@ func main() {
 		line = line[begin:]
 
 		var e cef.Event
-		if err := e.Unpack(line, opts...); err != nil {
+		if err := e.Unpack(string(line), opts...); err != nil {
 			log.Println("ERROR:", err, "in:", string(line))
 		}